Warning fixes.

+++ trunk/common/image.cpp
@@ -151,7 +151,7 @@ bool Image::FileLoad(lcMemFile& File)
 	unsigned char* Buffer = File.mBuffer + File.mPosition;
 	size_t BufferLength = File.mFileSize - File.mPosition;
 
-	if (!Image.loadFromData(Buffer, BufferLength))
+	if (!Image.loadFromData(Buffer, (int)BufferLength))
 		return false;
 
 	CopyFromQImage(Image, *this);
